Try to fix recursive dependency injection
